Changed styles in header and nav bar
