+++ src/utils/Projects.js
@@ -1,19 +1,17 @@
-import React from 'react';
-import {Link} from 'react-router-dom';
 
-const Students = props => {
+import React from 'react';
+
+const Projects = pr => {
     return (
-      <div className="student-list">
-        {props.students.map(student => (
-          <Link to={`/projectlist/${student.studentId}`}>
-          <div className="student" key={student.studentId}>
-            <h2>{student.name}</h2>
-            <p>{student.email}</p>
+      <div className="project-list">
+        {pr.projects.map(project => (
+          <div className="project" key={project.id}>
+            <h2>{project.title}</h2>
+            <p>{project.notes}</p>
           </div>
-          </Link>
         ))}
       </div>
     );
   };
 
-export default Students;
+export default Projects;
